requirements.txt
aiodns==2.0.0
asynctest==0.13.0
base58==1.0.3
bitarray==1.0.1
coverage>=5.0.2
Events==0.3
lz4==2.2.1
neo3crypto==0.2
neo3vm>=0.5.1
neo3vm-stubs>=0.5.1
mmh3==2.5.1
mypy>=0.782
mypy-extensions==0.4.3
netaddr>=0.8.0
orjson>=3.4.6
jsonschema>=3.2.0
Sphinx==3.5.2
sphinx-autodoc-typehints==1.11.1
pycodestyle==2.5.0
pybiginteger>=1.2.3
pybiginteger-stubs>=1.2.3
docutils~=0.17
setuptools~=54.2.0
pip~=21.0.1
pycryptodome~=3.10.1